Move files to src/main/java
